ddev_backend/requirement.txt
django
psycopg2
djangorestframework
dj-rest-auth